dropdown проверки, скролл карточек, перемещение по клику в pane, навигация с клавиатуры

+++ src/components/CardTaskItem.vue
@@ -1,10 +1,10 @@
 <template>
-    <div class="card-container-items-item">
-        <label class="card-container-items-item-text">
-            <input @change="sendComplete" type="checkbox" v-model="complete" :disabled="complete" v-if="!taskInfo.edit">
+    <div class="card-container-items-item" tabindex="0">
+        <div class="card-container-items-item-text" @click="sendCompleteMode">
+            <input type="checkbox" v-model="complete" :disabled="complete" v-if="!taskInfo.edit">
             <span v-if="!taskInfo.edit">{{ text }}</span>
-            <textarea placeholder="Текст задачи" type="text" v-if="taskInfo.edit" v-model="text"></textarea>
-        </label>
+            <textarea v-on:keyup.ctrl.enter="addItem" placeholder="Текст задачи" type="text" v-if="taskInfo.edit" v-model="text" ref="textDOM"></textarea>
+        </div>
         <div class="card-container-items-item-edit">
             <div class="card-container-items-item-edit-date">
                 <span v-if="taskInfo.timeComplete != '00:00' && taskInfo.timeComplete != ''" class="completed">{{ taskInfo.dateComplete }} / {{ taskInfo.timeComplete }}</span>
@@ -12,9 +12,23 @@
                 <svg><use xlink:href="../assets/main.svg#icon_calendar"></use></svg>
             </div>
             <div class="card-container-items-item-edit-links">
-                <svg v-if="!taskInfo.edit" @click="editItem"><use xlink:href="../assets/main.svg#icon_pencil"></use></svg>
-                <svg v-if="!taskInfo.edit" @click="deleteItem"><use xlink:href="../assets/main.svg#icon_close"></use></svg>
-                <span v-if="taskInfo.edit" @click="addItem" data-label="checkbox"></span>
+                <button v-if="!taskInfo.edit" @click="editItem"><svg ><use xlink:href="../assets/main.svg#icon_pencil"></use></svg></button>
+                <button v-if="!taskInfo.edit" @click="deleteMode"><svg ><use xlink:href="../assets/main.svg#icon_close"></use></svg></button>
+                <button v-if="taskInfo.edit" @click="addItem" aria-label="checkbox"></button>
+                <div class="card-container-items-item-edit-links-dropdown" v-if="isDelete">
+                    <h5>Удалить задачу?</h5>
+                    <div class="card-container-items-item-edit-links-dropdown-content">
+                        <button @click="deleteItem('yes')">Да</button>
+                        <button @click="deleteItem('no')">Нет</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <div class="card-container-items-item-dropdown" v-if="isComplete">
+            <h5>Выполнить задачу?</h5>
+            <div class="card-container-items-item-dropdown-content">
+                <button @click="sendComplete('yes')">Да</button>
+                <button @click="sendComplete('no')">Нет</button>
             </div>
         </div>
     </div>
@@ -26,7 +40,20 @@ export default {
     data() {
         return {
             text: this.taskInfo.text,
-            complete: this.taskInfo.complete
+            complete: this.taskInfo.complete,
+            isEdit: false,
+            isDelete: false,
+            isComplete: false
+        }
+    },
+    updated() {
+        if (this.taskInfo.edit && !this.isEdit) { // клацнули на edit
+            this.isEdit = true;
+            this.$refs.textDOM.focus(); // ставим focus на title, когда EDIT MODE
+        }
+        if (!this.taskInfo.edit && this.isEdit) {
+            this.isEdit = false;
+            this.$el.focus(); // ставим фокус на карточку
         }
     },
     methods: {
@@ -41,11 +68,31 @@ export default {
         editItem() {
             this.$emit("updateCardItem", "task", "editItem", this.cardId, "", this.text, this.taskInfo.id, 0); // передается edit true / false
         },
-        deleteItem() {
-            this.$emit("updateCardItem", "task", "deleteItem", this.cardId, "", this.text, this.taskInfo.id, 0); // передается delete mode
+        deleteMode() {
+            this.isDelete = !this.isDelete;
         },
-        sendComplete() {
-            this.$emit("updateCardItem", "task", "sendComplete", this.cardId, "", this.text, this.taskInfo.id, this.complete); // complete true/false
+        deleteItem(mode) {
+            if (mode == "yes") {
+                this.$emit("updateCardItem", "task", "deleteItem", this.cardId, "", this.text, this.taskInfo.id, 0); // передается delete mode
+            }
+            if (mode == "no") {
+                this.isDelete = false;
+            }
+        },
+        sendCompleteMode() {
+            if (!this.complete && !this.taskInfo.edit) {
+                this.isComplete = !this.isComplete; // если пункт еще не выполнен || и он не редактируется
+            }
+        },
+        sendComplete(mode) {
+            if (mode == "yes") {
+                this.complete = true;
+                this.$emit("updateCardItem", "task", "sendComplete", this.cardId, "", this.text, this.taskInfo.id, this.complete); // complete true/false
+                this.isComplete = false;
+            }
+            if (mode == "no") {
+                this.isComplete = false;
+            }
         }
     }
 }
